refactor: rely on ghq for repo management
this simplifies the code base further and farm is now basically a thin
wrapper around ghq. The first implementation had fallback-options if ghq
is not available but I removed that due to more complex testing. If you
don't want to be using ghq you can just implement a dummy function that
corresponds to the same calls. It should be fairly simple.